multibot_streamlit/requirements.txt
openai
streamlit
python-dotenv
google-cloud-aiplatform
